Move <ProcessingStatusSummary /> to common component
