+++ app/javascript/components/common/index.ts
@@ -11,3 +11,4 @@ export { Avatar } from './Avatar'
 export { Modal } from './Modal'
 export { ProminentLink } from './ProminentLink'
 export { FormButton } from './FormButton'
+export { ProcessingStatusSummary } from './ProcessingStatusSummary'
