+++ app/javascript/components/track/IterationSummary.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useRef } from 'react'
 import { fromNow } from '../../utils/time'
 import { SubmissionMethodIcon } from './iteration-summary/SubmissionMethodIcon'
-import { ProcessingStatusSummary } from './iteration-summary/ProcessingStatusSummary'
+import { ProcessingStatusSummary } from '../common/ProcessingStatusSummary'
 import { AnalysisStatusSummary } from './iteration-summary/AnalysisStatusSummary'
 import { IterationChannel } from '../../channels/iterationChannel'
 import { Iteration } from '../types'
